OnefinProject/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
Django==3.2.1
djangorestframework==3.12.4
idna==2.10
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wincertstore==0.2